add config to init
